Move season switch into DevUtils module
